simplifying NETWORK arg calls

+++ entrypoint-node.sh
@@ -5,10 +5,11 @@ function cleanup()
 }
 
 DEFAULT_NODE_OPTIONS=--max-old-space-size=12288
+NETWORK=${NETWORK:-mainnet}
 
 if [[ "$@" != *"--node-config"* ]]
 then
-    NODE_CONFIG="--node-config ./cardano-world/docs/environments/${NETWORK:-mainnet}/config.json"
+    NODE_CONFIG="--node-config ./cardano-world/docs/environments/${NETWORK}/config.json"
 fi
 if [[ "$@" != *"--node-socket"* ]]
 then
@@ -23,19 +24,18 @@ if [ $ogmios_status -ne 0 ]; then
   exit $ogmios_status
 fi
 
-NODE_ENV=${NODE_ENV:-production} NETWORK=${NETWORK:-mainnet} NODE_OPTIONS="${NODE_OPTIONS:-$DEFAULT_NODE_OPTIONS}" npm start &
+NODE_ENV=${NODE_ENV:-production} NETWORK=${NETWORK} NODE_OPTIONS="${NODE_OPTIONS:-$DEFAULT_NODE_OPTIONS}" npm start &
 
 DB_FILE=/db/protocolMagicId
-
-if [ ! -f "$DB_FILE" ]; then
+if [ "${NETWORK}" == "mainnet" ] && [ ! -f "$DB_FILE" ]; then
     curl -o - https://downloads.csnapshots.io/snapshots/mainnet/$(curl -s https://downloads.csnapshots.io/snapshots/mainnet/mainnet-db-snapshot.json| jq -r .[].file_name ) | lz4 -c -d - | tar -x -C /
 fi
 
 trap cleanup SIGINT SIGTERM SIGKILL
 
 cardano-node run +RTS -N -RTS \
-    --config ./cardano-world/docs/environments/${NETWORK:-mainnet}/config.json \
-    --topology ./cardano-world/docs/environments/${NETWORK:-mainnet}/topology.json \
+    --config ./cardano-world/docs/environments/${NETWORK}/config.json \
+    --topology ./cardano-world/docs/environments/${NETWORK}/topology.json \
     --database-path /db \
     --port 3000 \
     --host-addr 0.0.0.0 \
